+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 #Release history
-##0.2.0
+##0.4.0
+* Added the infoText function
+* Added 2 new types: info and notice
+
+##0.3.0
 * Added the dot function
 * Added colortypes
 * Added the verbosepointer
+
+##0.2.0
 * Moved the source from `app` to `src`
+* Added license
 
 ##0.1.2
 Added documentation